requirements.txt
# Core dependencies
click==8.1.7
requests==2.31.0
colorama==0.4.6
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pylint==3.0.3
flake8==6.1.0
mypy==1.7.1
black==23.12.1
pre-commit==3.6.0
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14
